refactor(app): Move ipfs, subgraph & connectors libs
