+++ packages/web-app/components/list/context/AgreementListProvider.tsx
@@ -1,15 +1,11 @@
 import { ReactNode, useState } from "react";
 import { AgreementListContext } from "./AgreementListContext";
 import { Agreement } from "./types";
-import { ApolloClient, InMemoryCache, ApolloProvider } from "@apollo/client";
+import { ApolloProvider } from "@apollo/client";
+import { client } from "../../../lib/subgraph";
 
 export const AgreementListProvider = ({ children }: { children: ReactNode }) => {
 	const [agreements, setAgreements] = useState<Agreement[]>([]);
-
-	const client = new ApolloClient({
-		uri: process.env.NEXT_PUBLIC_GRAPH_API_URL,
-		cache: new InMemoryCache(),
-	});
 
 	const provider = {
 		agreements,
